data/news_ai_server/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
boto3

pandas
numpy
pymysql